+++ example/lib/main.dart
@@ -24,7 +24,15 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _selectedIndex = 0;
+  final _bottomBarController = BottomBarWithSheetController(initialIndex: 0);
+
+  @override
+  void initState() {
+    _bottomBarController.itemsStream.listen((i) {
+      print('Index $i is selcted');
+    });
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -41,7 +49,24 @@ class _MyHomePageState extends State<MyHomePage> {
         ),
       ),
       bottomNavigationBar: BottomBarWithSheet(
-        selectedIndex: _selectedIndex,
+        controller: _bottomBarController,
+        bottomBarTheme: BottomBarTheme(
+          mainButtonPosition: MainButtonPosition.middle,
+          decoration: BoxDecoration(
+            color: Colors.white,
+            borderRadius: BorderRadius.vertical(top: Radius.circular(25)),
+          ),
+          itemIconColor: Colors.grey,
+          itemTextStyle: TextStyle(
+            color: Colors.grey,
+            fontSize: 10.0,
+          ),
+          selectedItemTextStyle: TextStyle(
+            color: Colors.blue,
+            fontSize: 10.0,
+          ),
+        ),
+        onSelectItem: (index) => print(index),
         sheetChild: Center(
           child: Text(
             "Another content",
@@ -52,7 +77,6 @@ class _MyHomePageState extends State<MyHomePage> {
             ),
           ),
         ),
-        onSelectItem: (index) => setState(() => _selectedIndex = index),
         items: [
           BottomBarWithSheetItem(icon: Icons.people),
           BottomBarWithSheetItem(icon: Icons.shopping_cart),
@@ -63,79 +87,3 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 }
-
-// import 'package:bottom_bar_with_sheet/bottom_bar_with_sheet.dart';
-// import 'package:flutter/material.dart';
-
-// void main() => runApp(MyApp());
-
-// class MyApp extends StatelessWidget {
-//   @override
-//   Widget build(BuildContext context) {
-//     return MaterialApp(
-//       debugShowCheckedModeBanner: false,
-//       home: MyHomePage(),
-//     );
-//   }
-// }
-
-// class MyHomePage extends StatefulWidget {
-//   @override
-//   _MyHomePageState createState() => _MyHomePageState();
-// }
-
-// class _MyHomePageState extends State<MyHomePage> {
-//   int _selectedIndex = 0;
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       backgroundColor: Colors.blue,
-//       body: Center(
-//         child: Text(
-//           "Place for your content",
-//           style: TextStyle(
-//             color: Colors.white,
-//             fontSize: 20,
-//             fontWeight: FontWeight.w900,
-//           ),
-//         ),
-//       ),
-//       bottomNavigationBar: BottomBarWithSheet(
-//         selectedIndex: _selectedIndex,
-//         sheetChild: Center(
-//           child: Text(
-//             "Another content",
-//             style: TextStyle(
-//               color: Colors.grey[600],
-//               fontSize: 20,
-//               fontWeight: FontWeight.w900,
-//             ),
-//           ),
-//         ),
-//         bottomBarTheme: BottomBarTheme(
-//           mainButtonPosition: MainButtonPosition.middle,
-//           selectedItemBackgroundColor: Colors.white,
-//           selectedItemIconColor: Colors.blue,
-//           selectedItemIconSize: 20,
-//         ),
-//         mainActionButtonTheme: MainActionButtonTheme(
-//           size: 55,
-//           icon: Icon(
-//             Icons.add,
-//             color: Colors.white,
-//             size: 35,
-//           ),
-//         ),
-//         onSelectItem: (index) => setState(() => _selectedIndex = index),
-//         items: [
-//           BottomBarWithSheetItem(icon: Icons.people),
-//           BottomBarWithSheetItem(icon: Icons.shopping_cart),
-//           BottomBarWithSheetItem(icon: Icons.settings),
-//           BottomBarWithSheetItem(icon: Icons.favorite),
-//         ],
-//       ),
-//     );
-//   }
-// }
-//