+++ lib/src/theme/theme.dart
@@ -1,0 +1,2 @@
+export 'bottom_bar_with_sheet_theme.dart';
+export 'main_action_button_theme.dart';